% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrk.R
\name{calculate_rho}
\alias{calculate_rho}
\title{Calculate scaling parameter so that the max value is no more than 10. This allows
for the Matern function to give a nice range of values and not just squishing 
everything down to 1.0.}
\usage{
calculate_rho(K, scale = 1)
}
\arguments{
\item{K}{(data.frame or matrix) A data frame or matrix of coordinates}

\item{scale}{(real) The scale value}
}
\value{
(real) scale * rho
}
\description{
Calculate scaling parameter so that the max value is no more than 10. This allows
for the Matern function to give a nice range of values and not just squishing 
everything down to 1.0.
}
\examples{
calculate_rho(data.frame(x=c(1, 2, 3), y=c(4, 5, 6)))
}
