% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_mat}
\alias{cor_mat}
\title{Make a correlation matrix with lower diagonal masked (for ease of readability)}
\usage{
cor_mat(df)
}
\arguments{
\item{df}{(data.frame) NEBCS data}
}
\value{
(matrix) Pearson correlation matrix
}
\description{
Make a correlation matrix with lower diagonal masked (for ease of readability)
}
\examples{
cor_mat(df)
}
