% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_na}
\alias{count_na}
\title{Returns number of rows with NA values in any of the cols supplied}
\usage{
count_na(df, cols = NULL)
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{cols}{(list) list of column names to look for NA values. Defaults to all columns.}
}
\value{
(number) Number of rows with NA values in any of the cols supplied
}
\description{
Returns number of rows with NA values in any of the cols supplied
}
\examples{
count_na(df, c('YEAR', 'CIGDUR'))
}
