% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrk.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance function}
\usage{
euc_dist(x1, x2)
}
\arguments{
\item{x1}{(tuple<real>) A tuple of real numbers}

\item{x2}{(tuple<real>) A tuple of real numbers}
}
\value{
(real) Euclidean distance between x1 and x2
}
\description{
Euclidean distance function
}
\examples{
dist(c(1, 2), c(3, 4))
}
