% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_na}
\alias{fill_na}
\title{Fill rows with NA with the mean or median of the column}
\usage{
fill_na(df, cols = NULL, method = "median")
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{cols}{(list) list of column names to look for NA values}

\item{method}{(str) (default "median") "mean" or "median"}
}
\value{
(data.frame) Same data as df, but with rows with NA values in any of the requested columns filled
}
\description{
Fill rows with NA with the mean or median of the column
}
\examples{
fill_na(df, c('YEAR', 'CIGDUR'), method="mean")
}
