% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_case_cntl}
\alias{get_case_cntl}
\title{Get a data frame with just the PID and CASE_CNTL values}
\usage{
get_case_cntl(df)
}
\arguments{
\item{df}{(data.frame) NEBCS data}
}
\value{
(data.frame) Data frame with just the PID and CASE_CNTL values
}
\description{
Get a data frame with just the PID and CASE_CNTL values
}
\examples{
get_case_cntl(df)
}
