% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{get_health_cols}
\alias{get_health_cols}
\title{Returns health column names}
\usage{
get_health_cols()
}
\value{
(array) Array of health column names
}
\description{
Returns health column names
}
\examples{
get_health_cols()
}
