% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutants.R
\name{get_pollutant_cols}
\alias{get_pollutant_cols}
\title{Returns pollutant columns}
\usage{
get_pollutant_cols()
}
\value{
(array) Array of pollutant column names
}
\description{
Returns pollutant columns
}
\examples{
get_pollutant_cols()
}
