% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutants.R
\name{group_pollutant_cols}
\alias{group_pollutant_cols}
\title{Returns named list of pollutant column names, grouped by pollutant type}
\usage{
group_pollutant_cols()
}
\value{
(list<array>) Named list of pollutant column names, grouped by pollutant type
}
\description{
Returns named list of pollutant column names, grouped by pollutant type
}
\examples{
group_pollutant_cols()
}
