% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_data}
\alias{load_data}
\title{Load the cleaned and jittered data. Optionally do various cleaning/data prep
tasks.
#' @export}
\usage{
load_data(
  google_id,
  dedup = FALSE,
  drop_missing_coords = FALSE,
  to_numeric = FALSE,
  N = 0
)
}
\arguments{
\item{google_id}{(str) Google file ID. File must be set to shareable via link.}

\item{dedup}{(bool) (Optional) Whether or not to remove duplicate rows (occurs when someone moves)}

\item{drop_missing_coords}{(bool) (Optional) Whether or not to drop rows with missing X/Y coordinates}

\item{to_numeric}{(bool) (Optional) Whether or not to convert character columns (gender) to numeric}

\item{N}{(int) (Optional) A positive integer indicating the years to keep for each person}
}
\value{
(data.frame) A data frame of people
}
\description{
Load the cleaned and jittered data. Optionally do various cleaning/data prep
tasks.
#' @export
}
\examples{
load_data('abc123', dedup=TRUE, drop_missing_coords=FALSE, ...)
}
