% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrk.R
\name{make_dist_mat}
\alias{make_dist_mat}
\title{Given a data frame or matrix where the first column corresponds to the x-value
and the second column corresponds to the y-value (or vice-versa), returns a
matrix where the i-j'th coordinate is the Euclidean distance between the
i'th and j'th rows.}
\usage{
make_dist_mat(K)
}
\arguments{
\item{K}{(data.frame or matrix) A data frame or matrix of coordinates}
}
\value{
(matrix) A matrix of Euclidean distances
}
\description{
Given a data frame or matrix where the first column corresponds to the x-value
and the second column corresponds to the y-value (or vice-versa), returns a
matrix where the i-j'th coordinate is the Euclidean distance between the
i'th and j'th rows.
}
\examples{
K = data.frame(x=c(1, 2, 3), y=c(4, 5, 6))
make_dist_mat(K)
}
