% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrk.R
\name{make_knots}
\alias{make_knots}
\title{Choose knots according to a swapping algorithm. Uses cover.design from the
fields package.}
\usage{
make_knots(df, N_k, ...)
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{N_k}{(int) Number of knots to choose}

\item{...}{Additional named arguments to pass to cover.design}
}
\value{
(data.frame) Data frame with columns x and y
}
\description{
Choose knots according to a swapping algorithm. Uses cover.design from the
fields package.
}
\examples{
get_case_cntl(df)
}
