% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrk.R
\name{matern}
\alias{matern}
\title{Matern function}
\usage{
matern(d)
}
\arguments{
\item{d}{(real) A real number}
}
\value{
(real) Matern function evaluated at d
}
\description{
Matern function
}
\examples{
matern(-1.23)
}
