% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onehot}
\alias{onehot}
\title{One-hot categorical columns}
\usage{
onehot(df, cols, drop_orig = TRUE)
}
\arguments{
\item{df}{(data.frame) Data}

\item{cols}{(array) Array of names of columns to one-hot}

\item{drop_orig}{(bool) (default=TRUE), Whether or not drop the column and
only retain the newly created one-hotted columns. If FALSE, the original
columns will be included as well as the one-hotted columns.}
}
\value{
(data.frame) Same data as df, additional columns for all columns
to one-hot. New column names are of the form `old-col-name_col-value`
}
\description{
One-hot categorical columns
}
\examples{
onehot(df, cols=c('col1', 'col2'))
}
