% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{powerset}
\alias{powerset}
\title{Calculate the powerset of a list or similar object.
Source: https://rdrr.io/cran/rje/src/R/powerSet.R}
\usage{
powerset(x)
}
\arguments{
\item{x}{(set, list, array) 1-dimensional object}
}
\value{
(set, list, array) Object of the same type as parameter s
}
\description{
Calculate the powerset of a list or similar object.
Source: https://rdrr.io/cran/rje/src/R/powerSet.R
}
\examples{
powerset(c(1, 2, 3))
}
