% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{predict_spca}
\alias{predict_spca}
\title{Reduce dimensions of new data using the model from spca().
See original paper (https://tibshirani.su.domains/ftp/spca.pdf) for details.}
\usage{
predict_spca(df, X_cols, y_col = "CASE_CNTL", model)
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{X_cols}{(array<character>) reduced columns found from spca(). Should use spca()$cols.}

\item{y_col}{(character) (default='CASE_CNTL') column name for the response variable}

\item{model}{(Object) best model found from spca(). Should use spca()$model.}
}
\value{
(array<numeric>) Pollutants reduced to one dimension using results from spca().
}
\description{
Reduce dimensions of new data using the model from spca().
See original paper (https://tibshirani.su.domains/ftp/spca.pdf) for details.
}
\examples{
res = spca(...)
predict_spca(df, X_cols=res$cols, model=res$model)
}
