% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_na}
\alias{remove_na}
\title{Remove rows with NA values in any of the requested columns}
\usage{
remove_na(df, cols = NULL)
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{cols}{(list) list of column names to look for NA values}
}
\value{
(data.frame) Same data as df, but with rows with NA values in any of the requested columns removed
}
\description{
Remove rows with NA values in any of the requested columns
}
\examples{
remove_na(df, c('YEAR', 'CIGDUR'))
}
