% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{train_test_split}
\alias{train_test_split}
\title{Train-test split by person, so that no person shows up in both the training and test set.
Note that test_size refers to the number/percentage of _people_, not of _rows_.}
\usage{
train_test_split(df, test_size, combine = FALSE)
}
\arguments{
\item{df}{(data.frame) NEBCS data}

\item{test_size}{(number) Float between 0 and 1 indicating a percentage of the data to use
as the test set, or integer indicating the number of rows to use as the test set.}

\item{combine}{(bool) (default=FALSE) Whether or not to combine the training and test sets with test
set marked as `validation == 1`.}
}
\value{
(list<data.frame>) Training set, test set, optionally combined set.
}
\description{
Train-test split by person, so that no person shows up in both the training and test set.
Note that test_size refers to the number/percentage of _people_, not of _rows_.
}
\examples{
tts_df_list = get_case_cntl(df)
tts_df_list$train
tts_df_list$test
}
