% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{fill_missing_cigdur}
\alias{fill_missing_cigdur}
\title{Fill missing smoking values using the medians for each CIG_CAT}
\usage{
fill_missing_cigdur(df)
}
\arguments{
\item{df}{(data.frame) NEBCS data}
}
\value{
df (data.frame) A data frame of people
}
\description{
Fill missing smoking values using the medians for each CIG_CAT
}
\examples{
fill_missing_cigdur(df)
}
