% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_data}
\alias{load_data}
\title{Load the cleaned and jittered data. Optionally do various cleaning/data prep
tasks.}
\usage{
load_data(
  path,
  dedup = FALSE,
  drop_missing_coords = FALSE,
  to_numeric = FALSE,
  N = 0
)
}
\arguments{
\item{path}{(str) path to data file}

\item{dedup}{(bool) (Optional) Whether or not to remove duplicate rows (occurs when someone moves)}

\item{drop_missing_coords}{(bool) (Optional) Whether or not to drop rows with missing X/Y coordinates}

\item{to_numeric}{(bool) (Optional) Whether or not to convert character columns (gender) to numeric}

\item{N}{(int) (Optional) A positive integer indicating the years to keep for each person}
}
\value{
df (data.frame) A data frame of people
}
\description{
Load the cleaned and jittered data. Optionally do various cleaning/data prep
tasks.
}
\examples{
load_data('my_data.csv', dedup=TRUE, drop_missing_coords=FALSE, ...)
}
